+++ docs/user_documentation/code_example.md
@@ -2,6 +2,6 @@
 
 !!! note "About the example"
     Below you see the rendered version of the example module
-    `src\example_google.py`
+    `super_repo\example_google.py`
 
-::: src.example_google
+::: super_repo.example_google
